First attempt at creating a database
Today I tried to design the database. But its difficult to wrap my head arround the dynamics of the system. I was also testing out a database creating tool called, "DB Browser for SQLite". It works well, however, I predict that it will not provide the specifications we may need in the future. (i.e: specifying the format of data input)

The following links were used as references:

http://www.c-jump.com/bcc/common/Talk/SQL/SQL_50_Joins/SQL_50_Joins.html?fbclid=IwAR1D5_9KkDeRtMgrbnHUO93UkkUapplMtRL3Bw9APLsWBiWAwWYQL1RmjfE 

https://dba.stackexchange.com/questions/89801/database-design-for-attendance-feature 

https://creately.com/diagram/example/i9h4wnzc1/ER%20diagram%20for%20Student%20Attendance