Add files via upload
We tried to create the UI in Node.js, as it was the first language we were able to get a connection to out database with. Unfortunately, as Node is designed for the client side of the server, the UI proved virtually impossible to make because everything inside it is a thread and there is no sense of order in this unforgiving world. :(