Add files via upload
This was the attempt to make this project work in Java. It did not work. No matter what we tried, we were never able to actually establish a connection to the database. It was just really hard, so we embarked on a series of trying new languages to get this bad boy working.